Current version of Psi+ is 1.5.1548
It is based on:
* psi: ac2d3dc4
* plugins: d6f8fc4
* psimedia: fda9f7c
* resources: 247f5dd

+++ iris/src/irisnet/noncore/ice176.h
@@ -98,9 +98,11 @@ public:
     // one per local address.  you must set local addresses first.
     void setExternalAddresses(const QList<ExternalAddress> &addrs);
 
-    void setStunBindService(const QHostAddress &addr, int port); // REVIEW if we need both v4 and v6?
-    void setStunRelayUdpService(const QHostAddress &addr, int port, const QString &user, const QCA::SecureArray &pass);
-    void setStunRelayTcpService(const QHostAddress &addr, int port, const QString &user, const QCA::SecureArray &pass);
+    void setStunBindService(const QHostAddress &addr, quint16 port); // REVIEW if we need both v4 and v6?
+    void setStunRelayUdpService(const QHostAddress &addr, quint16 port, const QString &user,
+                                const QCA::SecureArray &pass);
+    void setStunRelayTcpService(const QHostAddress &addr, quint16 port, const QString &user,
+                                const QCA::SecureArray &pass);
 
     // these all start out enabled, but can be disabled for diagnostic
     //   purposes
